requirements.txt
flask
flask-cors
yt-dlp
google-genai
python-dotenv
firebase-admin
gunicorn